// public/apps/the_peoples_edh_bracket/the-peoples-edh-bracket-client/src/types/bindings/CardWithGlobalPoints.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CardLegality } from "./CardLegality";

export type CardWithGlobalPoints = { global_points: string, 
/**
 * A unique identifier for the mechanics / behavior of the card
 */
oracle_id: string, 
/**
 * The name of the card
 */
name: string, 
/**
 * An image uri for the card
 */
image_uri: string | null, 
/**
 * The legality of the card
 */
legality: CardLegality, };